Patch rgen_test.go so that it calls go mod to download vgu
The test TestFull in rgen_test.go dynamically creates a
go test file and a go.mod file in a tmp directory.

The test fails because the dependency github.com/vugu/vugu@v0.4.x
is not available, but is required (after the dependency updates).

The test has been updated so that it now downloads
github.com/vugu/vugu@master - note it always uses the master branch
and then calls go mod tidy to update the go.mod and go.sum.

The test then passes.